+++ src/main/java/com/company/financial/app/domain/model/dto/request/ClientFilterRequest.java
@@ -15,6 +15,6 @@ public class ClientFilterRequest implements Serializable {
     private String name;
     private String lastName;
     private String email;
-    private double moneyInvested;
+    private int moneyInvested;
     private String operatorComparator;
 }
