+++ src/main/java/com/company/financial/app/infrastructure/adapter/entity/ClientEntity.java
@@ -23,5 +23,5 @@ public class ClientEntity {
     private Long birthdate;
     private Long creationDate;
     private Long modificationDate;
-    private float moneyInvested;
+    private int moneyInvested;
 }
